requirements.txt
google-api-python-client == 2.82.0
google-api-core == 2.11.0
google-auth == 2.16.3
google-auth-oauthlib == 1.0.0